requirements.txt
langchain

langchain-google-genai

langgraph

python-dotenv

# In the terminal: pip install -r requirements.txt
